Refactoring GraphQL API from text schema example to work with namespaced GFS schemas
