+++ server/src/py/server.py
@@ -9,8 +9,6 @@ import logging
 
 import simplejson as json
 
-from functools import partial
-
 import asyncio
 import threading
 
@@ -18,41 +16,23 @@ from flask import Flask
 from flask_restful import Api
 from flask_cors import CORS, cross_origin
 from flask_swagger import swagger
-from flask import Response, request
+from flask import request
 from flask.views import View
 
-from graphql.type.schema import GraphQLSchema
-from graphql_server import (
-    HttpQueryError, 
-    default_format_error, 
-    encode_execution_results, 
-    json_encode, 
-    load_json_body, 
-    run_http_query
-)
-from graphql.backend import GraphQLCoreBackend
-
 from flask_graphql import GraphQLView
-from flask_graphql.render_graphiql import render_graphiql
 
 from gevent import pywsgi
 from geventwebsocket.handler import WebSocketHandler
-
-#
-# 
-# 
 
 from flask_socketio import SocketIO
 from flask_socketio import send, emit
 from flask_sockets import Sockets
 from graphql_ws.gevent import  GeventSubscriptionServer
 
-# from consumer import ReconnectingExampleConsumer
-from consumer import ExampleConsumer
+from gfs.api.graphql.resource.schema import GFSGQLSchemas
+from gfs.api.graphql.gql import GFSGQLView
 
-# 
-# 
-# 
+
 
 app = Flask(__name__)
 cors = CORS(app)
@@ -69,1427 +49,9 @@ socketio = SocketIO(app, cors_allowed_origins="*")
 sockets = Sockets(app)
 app.app_protocol = lambda environ_path_info: 'graphql-ws'
 
-# TODO: Remove 
-# mount_point = "/home/project"
-# mount_point = ""
-
 listen_addr = os.environ.get("LISTEN_ADDR", "0.0.0.0")
 listen_port = os.environ.get("LISTEN_PORT", "5000")
 
-from graphql_tools import build_executable_schema
-
-source_schema = """
-
-schema {
-  query: Query
-  mutation: Mutation
-  subscription: Subscription
-}
-
-type ComposeGQLType {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  name: String
-  compose: String
-  format: String
-}
-
-type CreateCompose {
-  ok: Boolean
-  error: String
-  instance: ComposeGQLType
-}
-
-type CreateQuery {
-  ok: Boolean
-  error: String
-  instance: QueryGQLType
-}
-
-type CreateTemplate {
-  ok: Boolean
-  error: String
-  instance: TemplateGQLType
-}
-
-type CreateType {
-  ok: Boolean
-  error: String
-  instance: TypeSchema
-}
-
-type DHCPService {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: Machine
-  description: String
-  name: String
-  subnets: [Subnet]
-  defaultLeaseTime: String
-  options: [DHCPServiceOption]
-  hosts: [MachineGroup]
-  maxLeaseTime: String
-  aliases: [MachineAliasGroup]
-  resolvers: [RemoteHostGroup]
-  deployed: [Ip]
-}
-
-type DHCPServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DHCPService
-}
-
-type DHCPServiceOption {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  value: String
-  name: String
-}
-
-type DHCPServiceOptionEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DHCPServiceOption
-}
-
-input DHCPServiceOptionInput {
-  id: String
-  uuid: String
-  tags: [String]
-  value: String!
-  name: String!
-}
-
-type DHCPServiceSubnet {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  rangeEnd: Ip
-  netmask: Ip
-  routes: [Ip]
-  network: Ip
-  nextServer: Ip
-  description: String
-  rangeStart: Ip
-  resolvers: [Ip]
-  routers: [Ip]
-  name: String
-  options: [DHCPServiceOption]
-}
-
-type DHCPServiceSubnetEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DHCPServiceSubnet
-}
-
-type DNSBINDService {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: Machine
-  description: String
-  name: String
-  subnets: [Subnet]
-  hosts: [MachineGroup]
-  aliases: [MachineAliasGroup]
-  resolvers: [RemoteHostGroup]
-  deployed: [Machine]
-}
-
-type DNSBINDServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DNSBINDService
-}
-
-type DNSMasqService {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: Machine
-  description: String
-  name: String
-  subnets: [Subnet]
-  hosts: [MachineGroup]
-  aliases: [MachineAliasGroup]
-  resolvers: [RemoteHostGroup]
-  deployed: [Ip]
-}
-
-type DNSMasqServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DNSMasqService
-}
-
-type DeleteCompose {
-  ok: Boolean
-  error: String
-}
-
-type DeleteQuery {
-  ok: Boolean
-  error: String
-}
-
-type DeleteTemplate {
-  ok: Boolean
-  error: String
-}
-
-type DeleteType {
-  ok: Boolean
-  error: String
-}
-
-type Deployable {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  isDeployed: String
-  target: Machine
-}
-
-type DeployableEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Deployable
-}
-
-type DeployedService {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: Machine
-  description: String
-  name: String
-  path: String
-  pcol: String
-  host: Machine
-  alias: MachineAlias
-  port: String
-}
-
-type DeployedServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DeployedService
-}
-
-type DeployedServiceGroup {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  members: [DeployedService]
-  description: String
-  name: String
-}
-
-type DeployedServiceGroupEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: DeployedServiceGroup
-}
-
-input DeployedServiceGroupInput {
-  id: String
-  uuid: String
-  tags: [String]
-  members: [DeployedServiceInput]
-  addMembers: [TypeRefInput]
-  description: String!
-  name: String!
-}
-
-input DeployedServiceInput {
-  id: String
-  uuid: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: MachineInput
-  setTarget: TypeRefInput
-  description: String
-  name: String!
-  path: String
-  pcol: String
-  host: MachineInput
-  setHost: TypeRefInput
-  alias: MachineAliasInput
-  setAlias: TypeRefInput
-  port: String!
-}
-
-type Extends {
-  name: String!
-  version: Int!
-}
-
-input ExtendsInput {
-  name: String!
-  version: Int = 1
-}
-
-type HTTPService {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-  isDeployed: String
-  target: Machine
-  description: String
-  name: String
-  services: [DeployedServiceGroup]
-  port: String
-}
-
-type HTTPServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: HTTPService
-}
-
-type Implements {
-  name: String!
-  version: Int!
-}
-
-input ImplementsInput {
-  name: String!
-  version: Int = 1
-}
-
-type Ip {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  description: String
-  name: String
-  address: String
-}
-
-type IpEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Ip
-}
-
-input IpInput {
-  id: String
-  uuid: String
-  tags: [String]
-  description: String
-  name: String!
-  address: String!
-}
-
-type Machine {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  interface: NetDevice
-  release: String
-  memory: String
-  machine: String
-  codename: String
-  interfaces: [NetDevice]
-  description: String
-  system: String
-  platform: String
-  vendor: String
-  version: String
-  processor: String
-  cores: String
-  kernel: String
-  cpus: String
-  arch: String
-  name: String
-}
-
-type MachineAlias {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  host: Machine
-  description: String
-  name: String
-}
-
-type MachineAliasEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: MachineAlias
-}
-
-type MachineAliasGroup {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  members: [MachineAlias]
-  description: String
-  name: String
-}
-
-type MachineAliasGroupEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: MachineAliasGroup
-}
-
-input MachineAliasGroupInput {
-  id: String
-  uuid: String
-  tags: [String]
-  members: [MachineAliasInput]
-  addMembers: [TypeRefInput]
-  description: String!
-  name: String!
-}
-
-input MachineAliasInput {
-  id: String
-  uuid: String
-  tags: [String]
-  host: MachineInput
-  setHost: TypeRefInput
-  description: String
-  name: String
-}
-
-type MachineEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Machine
-}
-
-type MachineGroup {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  members: [Machine]
-  description: String
-  name: String
-}
-
-type MachineGroupEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: MachineGroup
-}
-
-input MachineGroupInput {
-  id: String
-  uuid: String
-  tags: [String]
-  members: [MachineInput]
-  addMembers: [TypeRefInput]
-  description: String!
-  name: String!
-}
-
-input MachineInput {
-  id: String
-  uuid: String
-  tags: [String]
-  interface: NetDeviceInput
-  setInterface: TypeRefInput
-  release: String!
-  memory: String!
-  machine: String!
-  codename: String!
-  interfaces: [NetDeviceInput]
-  addInterfaces: [TypeRefInput]
-  description: String
-  system: String!
-  platform: String!
-  vendor: String!
-  version: String!
-  processor: String!
-  cores: String!
-  kernel: String!
-  cpus: String!
-  arch: String!
-  name: String!
-}
-
-type Mutation {
-  createQuery(filter: String, name: String!, query: String!): CreateQuery
-  updateQuery(filter: String, id: String!, name: String, query: String): UpdateQuery
-  deleteQuery(id: String!): DeleteQuery
-  createTemplate(format: String, name: String!, template: String!): CreateTemplate
-  updateTemplate(format: String, id: String!, name: String, template: String): UpdateTemplate
-  deleteTemplate(id: String!): DeleteTemplate
-  createType(extends: ExtendsInput, implements: [ImplementsInput], name: String!, properties: [PropertyInput]!, required: [String], type: String! = "object", version: Int = 1): CreateType
-  deleteType(name: String!, version: Int): DeleteType
-  createCompose(compose: String!, format: String, name: String!): CreateCompose
-  updateCompose(compose: String, format: String, id: String!, name: String): UpdateCompose
-  deleteCompose(id: String!): DeleteCompose
-  createDHCPServiceOption(name: String!, tags: [String], value: String!): createDHCPServiceOption
-  updateDHCPServiceOption(id: String!, name: String, tags: [String], value: String): updateDHCPServiceOption
-  deleteDHCPServiceOption(id: String!): deleteDHCPServiceOption
-  createIp(address: String!, description: String, name: String!, tags: [String]): createIp
-  updateIp(address: String, description: String, id: String!, name: String, tags: [String]): updateIp
-  deleteIp(id: String!): deleteIp
-  createSubnet(addResolvers: [TypeRefInput], addRouters: [TypeRefInput], addRoutes: [TypeRefInput], description: String, name: String, netmask: IpInput, network: IpInput, nextServer: IpInput, rangeEnd: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], routes: [IpInput], setNetmask: TypeRefInput, setNetwork: TypeRefInput, setNextServer: TypeRefInput, setRangeEnd: TypeRefInput, setRangeStart: TypeRefInput, tags: [String]): createSubnet
-  updateSubnet(addResolvers: [TypeRefInput], addRouters: [TypeRefInput], addRoutes: [TypeRefInput], delResolvers: [TypeRefInput], delRouters: [TypeRefInput], delRoutes: [TypeRefInput], description: String, id: String!, name: String, netmask: IpInput, network: IpInput, nextServer: IpInput, rangeEnd: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], routes: [IpInput], setNetmask: TypeRefInput, setNetwork: TypeRefInput, setNextServer: TypeRefInput, setRangeEnd: TypeRefInput, setRangeStart: TypeRefInput, tags: [String]): updateSubnet
-  deleteSubnet(id: String!): deleteSubnet
-  createDHCPServiceSubnet(addOptions: [TypeRefInput], addResolvers: [TypeRefInput], addRouters: [TypeRefInput], addRoutes: [TypeRefInput], description: String, name: String!, netmask: IpInput, network: IpInput, nextServer: IpInput, options: [DHCPServiceOptionInput], rangeEnd: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], routes: [IpInput], setNetmask: TypeRefInput, setNetwork: TypeRefInput, setNextServer: TypeRefInput, setRangeEnd: TypeRefInput, setRangeStart: TypeRefInput, tags: [String]): createDHCPServiceSubnet
-  updateDHCPServiceSubnet(addOptions: [TypeRefInput], addResolvers: [TypeRefInput], addRouters: [TypeRefInput], addRoutes: [TypeRefInput], delOptions: [TypeRefInput], delResolvers: [TypeRefInput], delRouters: [TypeRefInput], delRoutes: [TypeRefInput], description: String, id: String!, name: String, netmask: IpInput, network: IpInput, nextServer: IpInput, options: [DHCPServiceOptionInput], rangeEnd: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], routes: [IpInput], setNetmask: TypeRefInput, setNetwork: TypeRefInput, setNextServer: TypeRefInput, setRangeEnd: TypeRefInput, setRangeStart: TypeRefInput, tags: [String]): updateDHCPServiceSubnet
-  deleteDHCPServiceSubnet(id: String!): deleteDHCPServiceSubnet
-  createNetDevice(addAddresses: [TypeRefInput], addresses: [IpInput], description: String, hwaddr: String!, name: String!, tags: [String]): createNetDevice
-  updateNetDevice(addAddresses: [TypeRefInput], addresses: [IpInput], delAddresses: [TypeRefInput], description: String, hwaddr: String, id: String!, name: String, tags: [String]): updateNetDevice
-  deleteNetDevice(id: String!): deleteNetDevice
-  createMachine(addInterfaces: [TypeRefInput], arch: String!, codename: String!, cores: String!, cpus: String!, description: String, interface: NetDeviceInput, interfaces: [NetDeviceInput], kernel: String!, machine: String!, memory: String!, name: String!, platform: String!, processor: String!, release: String!, setInterface: TypeRefInput, system: String!, tags: [String], vendor: String!, version: String!): createMachine
-  updateMachine(addInterfaces: [TypeRefInput], arch: String, codename: String, cores: String, cpus: String, delInterfaces: [TypeRefInput], description: String, id: String!, interface: NetDeviceInput, interfaces: [NetDeviceInput], kernel: String, machine: String, memory: String, name: String, platform: String, processor: String, release: String, setInterface: TypeRefInput, system: String, tags: [String], vendor: String, version: String): updateMachine
-  deleteMachine(id: String!): deleteMachine
-  createDeployable(isDeployed: String, setTarget: TypeRefInput, tags: [String], target: MachineInput): createDeployable
-  updateDeployable(id: String!, isDeployed: String, setTarget: TypeRefInput, tags: [String], target: MachineInput): updateDeployable
-  deleteDeployable(id: String!): deleteDeployable
-  createMachineAlias(description: String, host: MachineInput, name: String, setHost: TypeRefInput, tags: [String]): createMachineAlias
-  updateMachineAlias(description: String, host: MachineInput, id: String!, name: String, setHost: TypeRefInput, tags: [String]): updateMachineAlias
-  deleteMachineAlias(id: String!): deleteMachineAlias
-  createMachineAliasGroup(addMembers: [TypeRefInput], description: String!, members: [MachineAliasInput], name: String!, tags: [String]): createMachineAliasGroup
-  updateMachineAliasGroup(addMembers: [TypeRefInput], delMembers: [TypeRefInput], description: String, id: String!, members: [MachineAliasInput], name: String, tags: [String]): updateMachineAliasGroup
-  deleteMachineAliasGroup(id: String!): deleteMachineAliasGroup
-  createMachineGroup(addMembers: [TypeRefInput], description: String!, members: [MachineInput], name: String!, tags: [String]): createMachineGroup
-  updateMachineGroup(addMembers: [TypeRefInput], delMembers: [TypeRefInput], description: String, id: String!, members: [MachineInput], name: String, tags: [String]): updateMachineGroup
-  deleteMachineGroup(id: String!): deleteMachineGroup
-  createRemoteHost(addAddresses: [TypeRefInput], addresses: [IpInput], description: String!, name: String!, tags: [String]): createRemoteHost
-  updateRemoteHost(addAddresses: [TypeRefInput], addresses: [IpInput], delAddresses: [TypeRefInput], description: String, id: String!, name: String, tags: [String]): updateRemoteHost
-  deleteRemoteHost(id: String!): deleteRemoteHost
-  createRemoteHostGroup(addMembers: [TypeRefInput], description: String!, members: [RemoteHostInput], name: String!, tags: [String]): createRemoteHostGroup
-  updateRemoteHostGroup(addMembers: [TypeRefInput], delMembers: [TypeRefInput], description: String, id: String!, members: [RemoteHostInput], name: String, tags: [String]): updateRemoteHostGroup
-  deleteRemoteHostGroup(id: String!): deleteRemoteHostGroup
-  createRunnable(isRunning: String, status: String, tags: [String]): createRunnable
-  updateRunnable(id: String!, isRunning: String, status: String, tags: [String]): updateRunnable
-  deleteRunnable(id: String!): deleteRunnable
-  createService(description: String, name: String!, tags: [String]): createService
-  updateService(description: String, id: String!, name: String, tags: [String]): updateService
-  deleteService(id: String!): deleteService
-  createDHCPService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addOptions: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], defaultLeaseTime: String!, deployed: [IpInput], description: String, hosts: [MachineGroupInput], isDeployed: String, isRunning: String, maxLeaseTime: String!, name: String!, options: [DHCPServiceOptionInput], resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): createDHCPService
-  updateDHCPService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addOptions: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], defaultLeaseTime: String, delAliases: [TypeRefInput], delDeployed: [TypeRefInput], delHosts: [TypeRefInput], delOptions: [TypeRefInput], delResolvers: [TypeRefInput], delSubnets: [TypeRefInput], deployed: [IpInput], description: String, hosts: [MachineGroupInput], id: String!, isDeployed: String, isRunning: String, maxLeaseTime: String, name: String, options: [DHCPServiceOptionInput], resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): updateDHCPService
-  deleteDHCPService(id: String!): deleteDHCPService
-  createDNSBINDService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], deployed: [MachineInput], description: String, hosts: [MachineGroupInput], isDeployed: String, isRunning: String, name: String!, resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): createDNSBINDService
-  updateDNSBINDService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], delAliases: [TypeRefInput], delDeployed: [TypeRefInput], delHosts: [TypeRefInput], delResolvers: [TypeRefInput], delSubnets: [TypeRefInput], deployed: [MachineInput], description: String, hosts: [MachineGroupInput], id: String!, isDeployed: String, isRunning: String, name: String, resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): updateDNSBINDService
-  deleteDNSBINDService(id: String!): deleteDNSBINDService
-  createDNSMasqService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], deployed: [IpInput], description: String, hosts: [MachineGroupInput], isDeployed: String, isRunning: String, name: String!, resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): createDNSMasqService
-  updateDNSMasqService(addAliases: [TypeRefInput], addDeployed: [TypeRefInput], addHosts: [TypeRefInput], addResolvers: [TypeRefInput], addSubnets: [TypeRefInput], aliases: [MachineAliasGroupInput], delAliases: [TypeRefInput], delDeployed: [TypeRefInput], delHosts: [TypeRefInput], delResolvers: [TypeRefInput], delSubnets: [TypeRefInput], deployed: [IpInput], description: String, hosts: [MachineGroupInput], id: String!, isDeployed: String, isRunning: String, name: String, resolvers: [RemoteHostGroupInput], setTarget: TypeRefInput, status: String, subnets: [SubnetInput], tags: [String], target: MachineInput): updateDNSMasqService
-  deleteDNSMasqService(id: String!): deleteDNSMasqService
-  createDeployedService(alias: MachineAliasInput, description: String, host: MachineInput, isDeployed: String, isRunning: String, name: String!, path: String, pcol: String, port: String!, setAlias: TypeRefInput, setHost: TypeRefInput, setTarget: TypeRefInput, status: String, tags: [String], target: MachineInput): createDeployedService
-  updateDeployedService(alias: MachineAliasInput, description: String, host: MachineInput, id: String!, isDeployed: String, isRunning: String, name: String, path: String, pcol: String, port: String, setAlias: TypeRefInput, setHost: TypeRefInput, setTarget: TypeRefInput, status: String, tags: [String], target: MachineInput): updateDeployedService
-  deleteDeployedService(id: String!): deleteDeployedService
-  createDeployedServiceGroup(addMembers: [TypeRefInput], description: String!, members: [DeployedServiceInput], name: String!, tags: [String]): createDeployedServiceGroup
-  updateDeployedServiceGroup(addMembers: [TypeRefInput], delMembers: [TypeRefInput], description: String, id: String!, members: [DeployedServiceInput], name: String, tags: [String]): updateDeployedServiceGroup
-  deleteDeployedServiceGroup(id: String!): deleteDeployedServiceGroup
-  createHTTPService(addServices: [TypeRefInput], description: String, isDeployed: String, isRunning: String, name: String!, port: String!, services: [DeployedServiceGroupInput], setTarget: TypeRefInput, status: String, tags: [String], target: MachineInput): createHTTPService
-  updateHTTPService(addServices: [TypeRefInput], delServices: [TypeRefInput], description: String, id: String!, isDeployed: String, isRunning: String, name: String, port: String, services: [DeployedServiceGroupInput], setTarget: TypeRefInput, status: String, tags: [String], target: MachineInput): updateHTTPService
-  deleteHTTPService(id: String!): deleteHTTPService
-  createServiceGroup(addMembers: [TypeRefInput], description: String!, members: [ServiceInput], name: String!, tags: [String]): createServiceGroup
-  updateServiceGroup(addMembers: [TypeRefInput], delMembers: [TypeRefInput], description: String, id: String!, members: [ServiceInput], name: String, tags: [String]): updateServiceGroup
-  deleteServiceGroup(id: String!): deleteServiceGroup
-}
-
-type NetDevice {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  hwaddr: String
-  description: String
-  name: String
-  addresses: [Ip]
-}
-
-type NetDeviceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: NetDevice
-}
-
-input NetDeviceInput {
-  id: String
-  uuid: String
-  tags: [String]
-  hwaddr: String!
-  description: String
-  name: String!
-  addresses: [IpInput]
-  addAddresses: [TypeRefInput]
-}
-
-type Property {
-  name: String
-  type: String
-  title: String
-  itemsref: String
-}
-
-input PropertyInput {
-  name: String!
-  type: String! = "string"
-  title: String
-  itemsref: String
-}
-
-type Query {
-  query(id: String!): QueryGQLType
-  queries(name: String): [QueryGQLType]
-  querys(name: String): [QueryGQLType]
-  template(id: String!): TemplateGQLType
-  templates(name: String): [TemplateGQLType]
-  type(name: String!, version: Int): TypeSchema
-  types(name: String, version: Int): [TypeSchema]
-  compose(name: String!): ComposeGQLType
-  composes(name: String): [ComposeGQLType]
-  DHCPServiceOptions(value: String, name: String, tags: [String]): [DHCPServiceOption]
-  DHCPServiceOption(id: String!): DHCPServiceOption
-  Ips(name: String, address: String, tags: [String]): [Ip]
-  Ip(id: String!): Ip
-  Subnets(rangeEnd: IpInput, netmask: IpInput, routes: [IpInput], network: IpInput, nextServer: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], name: String, tags: [String]): [Subnet]
-  Subnet(id: String!): Subnet
-  DHCPServiceSubnets(rangeEnd: IpInput, netmask: IpInput, routes: [IpInput], network: IpInput, nextServer: IpInput, rangeStart: IpInput, resolvers: [IpInput], routers: [IpInput], name: String, options: [DHCPServiceOptionInput], tags: [String]): [DHCPServiceSubnet]
-  DHCPServiceSubnet(id: String!): DHCPServiceSubnet
-  NetDevices(hwaddr: String, name: String, addresses: [IpInput], tags: [String]): [NetDevice]
-  NetDevice(id: String!): NetDevice
-  Machines(interface: NetDeviceInput, release: String, memory: String, machine: String, codename: String, interfaces: [NetDeviceInput], system: String, platform: String, vendor: String, version: String, processor: String, cores: String, kernel: String, cpus: String, arch: String, name: String, tags: [String]): [Machine]
-  Machine(id: String!): Machine
-  Deployables(isDeployed: String, target: MachineInput, tags: [String]): [Deployable]
-  Deployable(id: String!): Deployable
-  MachineAliass(host: MachineInput, name: String, tags: [String]): [MachineAlias]
-  MachineAlias(id: String!): MachineAlias
-  MachineAliasGroups(members: [MachineAliasInput], name: String, tags: [String]): [MachineAliasGroup]
-  MachineAliasGroup(id: String!): MachineAliasGroup
-  MachineGroups(members: [MachineInput], name: String, tags: [String]): [MachineGroup]
-  MachineGroup(id: String!): MachineGroup
-  RemoteHosts(name: String, addresses: [IpInput], tags: [String]): [RemoteHost]
-  RemoteHost(id: String!): RemoteHost
-  RemoteHostGroups(members: [RemoteHostInput], name: String, tags: [String]): [RemoteHostGroup]
-  RemoteHostGroup(id: String!): RemoteHostGroup
-  Runnables(status: String, isRunning: String, tags: [String]): [Runnable]
-  Runnable(id: String!): Runnable
-  Services(name: String, tags: [String]): [Service]
-  Service(id: String!): Service
-  DHCPServices(status: String, isRunning: String, isDeployed: String, target: MachineInput, name: String, subnets: [SubnetInput], defaultLeaseTime: String, options: [DHCPServiceOptionInput], hosts: [MachineGroupInput], maxLeaseTime: String, aliases: [MachineAliasGroupInput], resolvers: [RemoteHostGroupInput], deployed: [IpInput], tags: [String]): [DHCPService]
-  DHCPService(id: String!): DHCPService
-  DNSBINDServices(status: String, isRunning: String, isDeployed: String, target: MachineInput, name: String, subnets: [SubnetInput], hosts: [MachineGroupInput], aliases: [MachineAliasGroupInput], resolvers: [RemoteHostGroupInput], deployed: [MachineInput], tags: [String]): [DNSBINDService]
-  DNSBINDService(id: String!): DNSBINDService
-  DNSMasqServices(status: String, isRunning: String, isDeployed: String, target: MachineInput, name: String, subnets: [SubnetInput], hosts: [MachineGroupInput], aliases: [MachineAliasGroupInput], resolvers: [RemoteHostGroupInput], deployed: [IpInput], tags: [String]): [DNSMasqService]
-  DNSMasqService(id: String!): DNSMasqService
-  DeployedServices(status: String, isRunning: String, isDeployed: String, target: MachineInput, name: String, path: String, pcol: String, host: MachineInput, alias: MachineAliasInput, port: String, tags: [String]): [DeployedService]
-  DeployedService(id: String!): DeployedService
-  DeployedServiceGroups(members: [DeployedServiceInput], name: String, tags: [String]): [DeployedServiceGroup]
-  DeployedServiceGroup(id: String!): DeployedServiceGroup
-  HTTPServices(status: String, isRunning: String, isDeployed: String, target: MachineInput, name: String, services: [DeployedServiceGroupInput], port: String, tags: [String]): [HTTPService]
-  HTTPService(id: String!): HTTPService
-  ServiceGroups(members: [ServiceInput], name: String, tags: [String]): [ServiceGroup]
-  ServiceGroup(id: String!): ServiceGroup
-}
-
-type QueryGQLType {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  name: String
-  query: String
-  filter: String
-}
-
-type RemoteHost {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  description: String
-  name: String
-  addresses: [Ip]
-}
-
-type RemoteHostEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: RemoteHost
-}
-
-type RemoteHostGroup {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  members: [RemoteHost]
-  description: String
-  name: String
-}
-
-type RemoteHostGroupEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: RemoteHostGroup
-}
-
-input RemoteHostGroupInput {
-  id: String
-  uuid: String
-  tags: [String]
-  members: [RemoteHostInput]
-  addMembers: [TypeRefInput]
-  description: String!
-  name: String!
-}
-
-input RemoteHostInput {
-  id: String
-  uuid: String
-  tags: [String]
-  description: String!
-  name: String!
-  addresses: [IpInput]
-  addAddresses: [TypeRefInput]
-}
-
-type Runnable {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  status: String
-  isRunning: String
-}
-
-type RunnableEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Runnable
-}
-
-type Service {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  description: String
-  name: String
-}
-
-type ServiceEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Service
-}
-
-type ServiceGroup {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  members: [Service]
-  description: String
-  name: String
-}
-
-type ServiceGroupEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: ServiceGroup
-}
-
-input ServiceInput {
-  id: String
-  uuid: String
-  tags: [String]
-  description: String
-  name: String!
-}
-
-type Subnet {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  tags: [String]
-  rangeEnd: Ip
-  netmask: Ip
-  routes: [Ip]
-  network: Ip
-  nextServer: Ip
-  description: String
-  rangeStart: Ip
-  resolvers: [Ip]
-  routers: [Ip]
-  name: String
-}
-
-type SubnetEvent {
-  event: String
-  id: String
-  label: String
-  sourceid: String
-  sourcelabel: String
-  targetid: String
-  targetlabel: String
-  chain: [String]
-  node: Subnet
-}
-
-input SubnetInput {
-  id: String
-  uuid: String
-  tags: [String]
-  rangeEnd: IpInput
-  setRangeEnd: TypeRefInput
-  netmask: IpInput
-  setNetmask: TypeRefInput
-  routes: [IpInput]
-  addRoutes: [TypeRefInput]
-  network: IpInput
-  setNetwork: TypeRefInput
-  nextServer: IpInput
-  setNextServer: TypeRefInput
-  description: String
-  rangeStart: IpInput
-  setRangeStart: TypeRefInput
-  resolvers: [IpInput]
-  addResolvers: [TypeRefInput]
-  routers: [IpInput]
-  addRouters: [TypeRefInput]
-  name: String
-}
-
-type Subscription {
-  DHCPServiceOption(events: [String], chain: [String]): DHCPServiceOptionEvent
-  Ip(events: [String], chain: [String]): IpEvent
-  Subnet(events: [String], chain: [String]): SubnetEvent
-  DHCPServiceSubnet(events: [String], chain: [String]): DHCPServiceSubnetEvent
-  NetDevice(events: [String], chain: [String]): NetDeviceEvent
-  Machine(events: [String], chain: [String]): MachineEvent
-  Deployable(events: [String], chain: [String]): DeployableEvent
-  MachineAlias(events: [String], chain: [String]): MachineAliasEvent
-  MachineAliasGroup(events: [String], chain: [String]): MachineAliasGroupEvent
-  MachineGroup(events: [String], chain: [String]): MachineGroupEvent
-  RemoteHost(events: [String], chain: [String]): RemoteHostEvent
-  RemoteHostGroup(events: [String], chain: [String]): RemoteHostGroupEvent
-  Runnable(events: [String], chain: [String]): RunnableEvent
-  Service(events: [String], chain: [String]): ServiceEvent
-  DHCPService(events: [String], chain: [String]): DHCPServiceEvent
-  DNSBINDService(events: [String], chain: [String]): DNSBINDServiceEvent
-  DNSMasqService(events: [String], chain: [String]): DNSMasqServiceEvent
-  DeployedService(events: [String], chain: [String]): DeployedServiceEvent
-  DeployedServiceGroup(events: [String], chain: [String]): DeployedServiceGroupEvent
-  HTTPService(events: [String], chain: [String]): HTTPServiceEvent
-  ServiceGroup(events: [String], chain: [String]): ServiceGroupEvent
-}
-
-type TemplateGQLType {
-  id: String
-  uuid: String
-  created: String
-  modified: String
-  name: String
-  template: String
-  format: String
-}
-
-input TypeRefInput {
-  id: String
-}
-
-type TypeSchema {
-  id: String
-  name: String
-  type: String
-  version: Int
-  extends: Extends
-  implements: [Implements]
-  typelabel: String
-  required: [String]
-  properties: [Property]
-}
-
-type UpdateCompose {
-  ok: Boolean
-  error: String
-  instance: ComposeGQLType
-}
-
-type UpdateQuery {
-  ok: Boolean
-  error: String
-  instance: QueryGQLType
-}
-
-type UpdateTemplate {
-  ok: Boolean
-  error: String
-  instance: TemplateGQLType
-}
-
-type createDHCPService {
-  ok: Boolean
-  error: String
-  instance: DHCPService
-}
-
-type createDHCPServiceOption {
-  ok: Boolean
-  error: String
-  instance: DHCPServiceOption
-}
-
-type createDHCPServiceSubnet {
-  ok: Boolean
-  error: String
-  instance: DHCPServiceSubnet
-}
-
-type createDNSBINDService {
-  ok: Boolean
-  error: String
-  instance: DNSBINDService
-}
-
-type createDNSMasqService {
-  ok: Boolean
-  error: String
-  instance: DNSMasqService
-}
-
-type createDeployable {
-  ok: Boolean
-  error: String
-  instance: Deployable
-}
-
-type createDeployedService {
-  ok: Boolean
-  error: String
-  instance: DeployedService
-}
-
-type createDeployedServiceGroup {
-  ok: Boolean
-  error: String
-  instance: DeployedServiceGroup
-}
-
-type createHTTPService {
-  ok: Boolean
-  error: String
-  instance: HTTPService
-}
-
-type createIp {
-  ok: Boolean
-  error: String
-  instance: Ip
-}
-
-type createMachine {
-  ok: Boolean
-  error: String
-  instance: Machine
-}
-
-type createMachineAlias {
-  ok: Boolean
-  error: String
-  instance: MachineAlias
-}
-
-type createMachineAliasGroup {
-  ok: Boolean
-  error: String
-  instance: MachineAliasGroup
-}
-
-type createMachineGroup {
-  ok: Boolean
-  error: String
-  instance: MachineGroup
-}
-
-type createNetDevice {
-  ok: Boolean
-  error: String
-  instance: NetDevice
-}
-
-type createRemoteHost {
-  ok: Boolean
-  error: String
-  instance: RemoteHost
-}
-
-type createRemoteHostGroup {
-  ok: Boolean
-  error: String
-  instance: RemoteHostGroup
-}
-
-type createRunnable {
-  ok: Boolean
-  error: String
-  instance: Runnable
-}
-
-type createService {
-  ok: Boolean
-  error: String
-  instance: Service
-}
-
-type createServiceGroup {
-  ok: Boolean
-  error: String
-  instance: ServiceGroup
-}
-
-type createSubnet {
-  ok: Boolean
-  error: String
-  instance: Subnet
-}
-
-type deleteDHCPService {
-  ok: Boolean
-  error: String
-}
-
-type deleteDHCPServiceOption {
-  ok: Boolean
-  error: String
-}
-
-type deleteDHCPServiceSubnet {
-  ok: Boolean
-  error: String
-}
-
-type deleteDNSBINDService {
-  ok: Boolean
-  error: String
-}
-
-type deleteDNSMasqService {
-  ok: Boolean
-  error: String
-}
-
-type deleteDeployable {
-  ok: Boolean
-  error: String
-}
-
-type deleteDeployedService {
-  ok: Boolean
-  error: String
-}
-
-type deleteDeployedServiceGroup {
-  ok: Boolean
-  error: String
-}
-
-type deleteHTTPService {
-  ok: Boolean
-  error: String
-}
-
-type deleteIp {
-  ok: Boolean
-  error: String
-}
-
-type deleteMachine {
-  ok: Boolean
-  error: String
-}
-
-type deleteMachineAlias {
-  ok: Boolean
-  error: String
-}
-
-type deleteMachineAliasGroup {
-  ok: Boolean
-  error: String
-}
-
-type deleteMachineGroup {
-  ok: Boolean
-  error: String
-}
-
-type deleteNetDevice {
-  ok: Boolean
-  error: String
-}
-
-type deleteRemoteHost {
-  ok: Boolean
-  error: String
-}
-
-type deleteRemoteHostGroup {
-  ok: Boolean
-  error: String
-}
-
-type deleteRunnable {
-  ok: Boolean
-  error: String
-}
-
-type deleteService {
-  ok: Boolean
-  error: String
-}
-
-type deleteServiceGroup {
-  ok: Boolean
-  error: String
-}
-
-type deleteSubnet {
-  ok: Boolean
-  error: String
-}
-
-type updateDHCPService {
-  ok: Boolean
-  error: String
-  instance: DHCPService
-}
-
-type updateDHCPServiceOption {
-  ok: Boolean
-  error: String
-  instance: DHCPServiceOption
-}
-
-type updateDHCPServiceSubnet {
-  ok: Boolean
-  error: String
-  instance: DHCPServiceSubnet
-}
-
-type updateDNSBINDService {
-  ok: Boolean
-  error: String
-  instance: DNSBINDService
-}
-
-type updateDNSMasqService {
-  ok: Boolean
-  error: String
-  instance: DNSMasqService
-}
-
-type updateDeployable {
-  ok: Boolean
-  error: String
-  instance: Deployable
-}
-
-type updateDeployedService {
-  ok: Boolean
-  error: String
-  instance: DeployedService
-}
-
-type updateDeployedServiceGroup {
-  ok: Boolean
-  error: String
-  instance: DeployedServiceGroup
-}
-
-type updateHTTPService {
-  ok: Boolean
-  error: String
-  instance: HTTPService
-}
-
-type updateIp {
-  ok: Boolean
-  error: String
-  instance: Ip
-}
-
-type updateMachine {
-  ok: Boolean
-  error: String
-  instance: Machine
-}
-
-type updateMachineAlias {
-  ok: Boolean
-  error: String
-  instance: MachineAlias
-}
-
-type updateMachineAliasGroup {
-  ok: Boolean
-  error: String
-  instance: MachineAliasGroup
-}
-
-type updateMachineGroup {
-  ok: Boolean
-  error: String
-  instance: MachineGroup
-}
-
-type updateNetDevice {
-  ok: Boolean
-  error: String
-  instance: NetDevice
-}
-
-type updateRemoteHost {
-  ok: Boolean
-  error: String
-  instance: RemoteHost
-}
-
-type updateRemoteHostGroup {
-  ok: Boolean
-  error: String
-  instance: RemoteHostGroup
-}
-
-type updateRunnable {
-  ok: Boolean
-  error: String
-  instance: Runnable
-}
-
-type updateService {
-  ok: Boolean
-  error: String
-  instance: Service
-}
-
-type updateServiceGroup {
-  ok: Boolean
-  error: String
-  instance: ServiceGroup
-}
-
-type updateSubnet {
-  ok: Boolean
-  error: String
-  instance: Subnet
-}
-
-"""
-# 
-# Subscriber example
-# 
-# query = """
-#     subscription serviceevent {
-#         DHCPService {
-#             id
-#         }
-#     }
-# """
-# 
-# # Asynchronous request
-# loop = asyncio.get_event_loop()
-# loop.run_until_complete(
-#     client.subscribe(
-#         query=query, 
-#         handle=print
-#     )
-# )
-# 
-
-def subscription_dhcpservice_resolver(value, info, **args):
-    print("subscription_dhcpservice_resolver")
-
-resolvers = {
-  "Subscription": {
-      "DHCPService": subscription_dhcpservice_resolver
-  }
-}
-
-my_schema = build_executable_schema(source_schema, resolvers)
-
-# 
-# 
-# 
 
 
 @socketio.on('connect', namespace='/gfs1')
@@ -1507,164 +69,59 @@ def handle_message(message):
 
 
 @sockets.route('/subscriptions')
-def subscriptions_handler(ws):
+def echo_socket(ws):
     subscription_server = GeventSubscriptionServer(
-        my_schema
+        # GFSGQLSchema(
+        #     "gfs1", 
+        #     GFSGQLSchemas.instance()
+        # ) # GFSGQLSchemas.instance().schema("gfs1")
+        GFSGQLSchemas.instance().schema("gfs1")
     )
     subscription_server.handle(ws)
     return []
 
-@sockets.route('/graphql/subscriptions')
-def subscriptions_handler2(ws):
+@sockets.route('/<namespace>/graphql/subscriptions')
+def echo_socket2(ws, namespace):
     subscription_server = GeventSubscriptionServer(
-        my_schema
+        # GFSGQLSchema(
+        #     namespace, 
+        #     GFSGQLSchemas.instance()
+        # ) # GFSGQLSchemas.instance().schema(namespace)
+        GFSGQLSchemas.instance().schema(namespace)
     )
     subscription_server.handle(ws)
     return []
 
 
 
+# schemas = GFSGQLSchemas()
+# GFSGQLSchemas.instance(schemas)
 
-class CustomBackend(GraphQLCoreBackend):
-    def __init__(self, executor=None):
-        super().__init__(executor)
-        self.execute_params['allow_subscriptions'] = True
-
-
-
-class GQLView(View):
-
-    schema = None
-
-    methods = ['GET', 'POST', 'PUT', 'DELETE']
-
-    def __init__(self, **kwargs):
-        super(GQLView, self).__init__()
-        for key, value in kwargs.items():
-            if hasattr(self, key):
-                setattr(self, key, value)
-
-    def context(self):
-        return request
-
-    def schema(self):
-        return self.schema
-
-    def graphiql(self, params, result):
-        return render_graphiql(
-            params=params,
-            result=result
-        )
-
-    format_error = staticmethod(default_format_error)
-    encode = staticmethod(json_encode)
-
-    def dispatch_request(self):
-
-        try:
-
-            show_graphiql = request.method.lower() == 'get' and self.is_graphiql()
-
-            data = self.parse_body()
-
-            # opts = {}
-
-            opts = {
-                "MIDDLEWARE": [], 
-                "allow_subscriptions": True
-            }
-
-            execution_results, all_params = run_http_query(
-                self.schema,
-                request.method.lower(),
-                data,
-                query_data=request.args,
-                catch=show_graphiql,
-                context=self.context(),
-                middleware=[],
-                # backend=
-                backend=CustomBackend(),
-                **opts
-            )
-
-            result, status_code = encode_execution_results(
-                execution_results,
-                is_batch=isinstance(data, list),
-                format_error=self.format_error,
-                encode=partial(self.encode, pretty=True)
-            )
-
-            if show_graphiql:
-                return self.graphiql(
-                    params=all_params[0],
-                    result=result
-                )
-
-            return Response(
-                result,
-                status=status_code,
-                content_type='application/json'
-            )
-
-        except HttpQueryError as e:
-            return Response(
-                self.encode({
-                    'errors': [self.format_error(e)]
-                }),
-                status=e.status_code,
-                headers=e.headers,
-                content_type='application/json'
-            )
-
-    def parse_body(self):
-        if request.mimetype == 'application/graphql':
-            return {
-                'query': request.data.decode('utf8')
-            }
-
-        elif request.mimetype == 'application/json':
-            return load_json_body(
-                request.data.decode('utf8')
-            )
-
-        return {}
-
-    def is_graphiql(self):
-        return self.is_html()
-
-    def is_html(self):
-        best = request.accept_mimetypes \
-            .best_match(['application/json', 'text/html'])
-        return best == 'text/html' and \
-            request.accept_mimetypes[best] > \
-            request.accept_mimetypes['application/json']
-
-
-
-
-
-view_func = GQLView.as_view(
-    'graphql',
-    schema=my_schema
+view_func = GFSGQLView.as_view(
+    'graphql', 
+    namespace='gfs1', 
+    schemas=GFSGQLSchemas.instance()
 )
 app.add_url_rule(
-    '/graphql', 
+    '/<namespace>/graphql', 
     view_func=view_func
 )
 
 class GraphQLSchema(View):
 
-    def dispatch_request(self):
-        schema=my_schema
-        return str( schema )
+    def dispatch_request(self, namespace):
+        schemas = GFSGQLSchemas.instance()
+        return str( schemas.schema(namespace) )
 
 view_func2 = GraphQLSchema.as_view(
     'graphql2'
 )
 app.add_url_rule(
-    '/graphql/schema', 
+    '/<namespace>/graphql/schema', 
     view_func=view_func2
 )
+
+
 
 print(str(listen_addr))
 print(int(listen_port))
